requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
mysqlclient
python-dotenv
requests
drf-spectacular
django-cors-headers
django-filter
